+++ src/abstraction/i-field.ts
@@ -4,5 +4,6 @@ export interface IField {
   name?: string;
   alias?: string;
   symbol?: string;
+  ignore?: boolean;
   cardinality?: Cardinality;
 }