+++ src/reflection/filter.ts
@@ -1,3 +1,3 @@
 export function filter<T>(array: Array<T>, each: (item: T) => boolean): Array<T> {
-  return array.filter(each);
+  return Array.isArray(array) ? array.filter(each) : [];
 }